+++ src/lib/mongoBackend/CMakeLists.txt
@@ -43,7 +43,7 @@ SET (SOURCES
     mongoGetSubscriptions.cpp
     connectionOperations.cpp
     mongoSubCache.cpp
-    safeBsonGet.cpp
+    safeMongo.cpp
     dbFieldEncoding.cpp
     compoundResponses.cpp
 )
@@ -72,7 +72,7 @@ SET (HEADERS
     mongoGetSubscriptions.h
     connectionOperations.h
     mongoSubCache.h
-    safeBsonGet.h
+    safeMongo.h
     dbFieldEncoding.h
     compoundResponses.h
 )
