+++ src/lib/mongoBackend/MongoCommonRegister.cpp
@@ -41,7 +41,7 @@
 #include "mongoBackend/TriggeredSubscription.h"
 
 #include "mongoBackend/connectionOperations.h"
-#include "mongoBackend/safeBsonGet.h"
+#include "mongoBackend/safeMongo.h"
 #include "mongoBackend/dbConstants.h"
 
 using std::string;
