+++ src/lib/mongoBackend/mongoGetSubscriptions.cpp
@@ -31,7 +31,7 @@
 #include "mongoBackend/mongoGetSubscriptions.h"
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/connectionOperations.h"
-#include "mongoBackend/safeBsonGet.h"
+#include "mongoBackend/safeMongo.h"
 #include "mongoBackend/dbConstants.h"
 
 #include "mongo/client/dbclient.h"
