+++ src/lib/mongoBackend/mongoOntimeintervalOperations.cpp
@@ -32,7 +32,7 @@
 
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/connectionOperations.h"
-#include "mongoBackend/safeBsonGet.h"
+#include "mongoBackend/safeMongo.h"
 #include "mongoBackend/dbConstants.h"
 #include "mongoBackend/mongoOntimeintervalOperations.h"
