+++ src/lib/mongoBackend/mongoUnsubscribeContext.cpp
@@ -33,7 +33,7 @@
 #include "mongoBackend/connectionOperations.h"
 #include "mongoBackend/mongoUnsubscribeContext.h"
 #include "mongoBackend/mongoSubCache.h"
-#include "mongoBackend/safeBsonGet.h"
+#include "mongoBackend/safeMongo.h"
 #include "ngsi10/UnsubscribeContextRequest.h"
 #include "ngsi10/UnsubscribeContextResponse.h"
